Properly handle batchless requests

+++ app/models/batch/request_behaviour.rb
@@ -1,12 +1,12 @@
 module Batch::RequestBehaviour
   def self.included(base)
     base.class_eval do
-      has_one :batch_requests
-      has_one :batch, :through => :batch_requests
+      has_one :batch_request
+      has_one :batch, :through => :batch_request
 
       # For backwards compatibility
-      def batch_requests; [batch_request]; end
-      def batches; [batch_request]; end
+      def batch_requests; [batch_request].compact ; end
+      def batches; [batch_request].compact ; end
 
 
       # Identifies all requests that are not part of a batch.
@@ -19,11 +19,14 @@ module Batch::RequestBehaviour
     end
   end
 
+  def with_batch_id
+    yield batch.id if batch.present?
+  end
 
   def recycle_from_batch!
     ActiveRecord::Base.transaction do
       self.return_for_inbox!
-      self.batch_request.destroy
+      self.batch_request.destroy if self.batch_request.present?
       self.save!
     end
     #self.detach
